separate reusable the function on common file
